requirements.txt
fastapi
uvicorn
python-dotenv
requests
langchain
langchain-community
langchain-huggingface
sentence-transformers
pypdf
faiss-cpu